// src/interfaces/registros.ts
export interface Registros {
    id?: number;   
    imagenBase64: string;   
    nombre: string;         
    descripcion: string;    
    asignatura: string;  
    fecha: string;             
    profesor: string;      
    qrData: string;         
    rut: string;            
    correo: string;         
  }